refactor: extract header bar into separate file
