+++ Frontend/react-app/src/components/AppLayout/AppLayout.css
@@ -1,67 +1,8 @@
 .app-container {
     display: flex;
     flex-direction: column;
-    min-height: 100vh;
+    min-height: 100%;
     background-color: var(--enmGray);
-}
-
-.app-layout-list-title {
-    margin: 0;
-    font-size: clamp(2rem, 2vw, 1.8rem);
-    color: var(--white);
-    max-width: 100%;
-}
-
-.app-header {
-    background-color: #2c3e50;
-    color: white;
-    padding: 1rem;
-    width: 100%;
-    position: sticky;
-    top: 0;
-    z-index: 100;
-}
-
-.header-content {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    max-width: 1200px;
-    margin: 0 auto;
-    width: 100%;
-    gap: 1rem;
-    padding: 0 1rem;
-}
-
-.app-title {
-    color: var(--white);
-    font-size: 1.5rem;
-    margin: 0;
-    white-space: nowrap;
-}
-
-.search-container {
-    flex: 1;
-    max-width: 600px;
-    margin: 0 1rem;
-}
-
-.logout-button {
-    background: none;
-    border: none;
-    color: white;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    font-size: 1rem;
-    padding: 0.5rem;
-    border-radius: 4px;
-    white-space: nowrap;
-}
-
-.logout-button:hover {
-    background-color: rgba(255, 255, 255, 0.1);
 }
 
 .app-content-wrapper {
@@ -79,8 +20,8 @@
     background-color: #fff;
     box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
     padding: 20px;
-    min-height: calc(100vh - 68px);
-    overflow: hidden;
+    height: calc(100vh - 68px);
+        overflow-y: auto;
 }
 
 .content-wrapper {
@@ -105,20 +46,10 @@
     justify-content: center;
     cursor: pointer;
     box-shadow: 0 2px 10px rgba(0, 0, 0, 0.2);
-    z-index: 100;
+    z-index: 10000;
     border: 3px solid white;
     transition: all 0.3s ease;
     animation: pulse 2s infinite;
-}
-
-.bug-report-button:hover {
-    background-color: #d45a1f;
-    transform: scale(1.15) rotate(10deg);
-    box-shadow: 0 6px 20px rgba(0, 0, 0, 0.4);
-}
-
-.bug-report-button:active {
-    transform: scale(0.95);
 }
 
 @keyframes pulse {
@@ -135,174 +66,9 @@
     }
 }
 
-@media (max-width: 1200px) {
-
-    .header-content,
-    .app-content {
-        padding: 1rem;
-    }
-}
-
 @media (max-width: 768px) {
-    .header-content {
-        flex-direction: column;
-        gap: 0.5rem;
-    }
-
-    .search-container {
-        width: 100%;
-        margin: 0.5rem 0;
-    }
-
     .app-content {
         padding: 10px;
     }
 }
 
-.mobile-menu-button {
-    background: none;
-    border: none;
-    color: white;
-    font-size: 1.5rem;
-    cursor: pointer;
-    display: none;
-    padding: 0.5rem;
-}
-
-.mobile-menu {
-    position: fixed;
-    top: 0;
-    right: -300px;
-    width: 250px;
-    height: 100vh;
-    background-color: #2c3e50;
-    padding: 1rem;
-    z-index: 101;
-    transition: right 0.3s ease;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.mobile-menu.open {
-    right: 0;
-}
-
-.mobile-menu-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    z-index: 100;
-    display: none;
-}
-
-.mobile-menu.open+.mobile-menu-overlay {
-    display: block;
-}
-
-@media (max-width: 768px) {
-    .mobile-menu-button {
-        display: block;
-    }
-
-    .header-content .button:not(.mobile-menu-button) {
-        display: none;
-    }
-
-    .header-content {
-        flex-direction: row;
-        justify-content: space-between;
-        gap: 1rem;
-    }
-
-    .search-container {
-        order: 1;
-        flex: 1;
-        margin: 0;
-    }
-
-    .app-title {
-        order: 0;
-        font-size: 1.2rem;
-    }
-
-    .mobile-menu-button {
-        order: 2;
-    }
-}
-
-@media (max-width: 480px) {
-    .header-content {
-        padding: 0.5rem;
-    }
-
-    .app-title {
-        font-size: 1rem;
-    }
-}
-
-.mobile-menu-close {
-    position: absolute;
-    top: 1rem;
-    right: 1rem;
-    background: none;
-    border: none;
-    color: white;
-    font-size: 1.5rem;
-    cursor: pointer;
-    padding: 0.5rem;
-}
-
-.mobile-menu-content {
-    margin-top: 10px;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
-
-.mobile-menu {
-    position: fixed;
-    top: 0;
-    right: -300px;
-    width: 250px;
-    height: 100vh;
-    background-color: #2c3e50;
-    padding-top: 3rem;
-    z-index: 101;
-    transition: right 0.3s ease;
-    display: flex;
-    flex-direction: column;
-}
-
-.mobile-menu.open {
-    right: 0;
-    box-shadow: -2px 0 10px rgba(0, 0, 0, 0.2);
-}
-
-.mobile-menu-overlay {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.5);
-    z-index: 100;
-    opacity: 0;
-    transition: opacity 0.3s ease;
-    pointer-events: none;
-}
-
-.mobile-menu.open+.mobile-menu-overlay {
-    opacity: 1;
-    pointer-events: auto;
-}
-
-@media (max-width: 768px) {
-    .mobile-menu-button {
-        position: relative;
-        z-index: 102;
-    }
-}
