+++ Frontend/react-app/src/components/HeaderBar/HeaderBar.css
@@ -1,22 +1,122 @@
+
 .headbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1rem;
     background-color: #2c3e50;
     color: white;
+    padding: 1rem;
+    width: 100%;
     position: sticky;
     top: 0;
     z-index: 100;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 1rem;
 }
 
 .headbar-title {
+    color: white;
     font-size: 1.5rem;
-    font-weight: bold;
+    margin: 0;
+    white-space: nowrap;
 }
 
 .headbar-search {
     flex: 1;
-    max-width: 500px;
+    max-width: 600px;
     margin: 0 1rem;
+}
+
+.headbar-logout {
+    background: none;
+    border: none;
+    color: white;
+}
+
+/* Mobile menu styles */
+.mobile-menu-button {
+    background: none;
+    border: none;
+    color: white;
+    font-size: 1.5rem;
+    cursor: pointer;
+    padding: 0.5rem;
+    order: 2; /* Ensures it appears on the right */
+}
+
+.mobile-menu {
+    position: fixed;
+    top: 0;
+    right: -300px;
+    width: 250px;
+    height: 100vh;
+    background-color: #2c3e50;
+    padding-top: 3rem;
+    z-index: 101;
+    transition: right 0.3s ease;
+    display: flex;
+    flex-direction: column;
+}
+
+.mobile-menu.open {
+    right: 0;
+    box-shadow: -2px 0 10px rgba(0, 0, 0, 0.2);
+}
+
+.mobile-menu-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    z-index: 100;
+    opacity: 0;
+    transition: opacity 0.3s ease;
+    pointer-events: none;
+}
+
+.mobile-menu.open+.mobile-menu-overlay {
+    opacity: 1;
+    pointer-events: auto;
+}
+
+.mobile-menu-close {
+    position: absolute;
+    top: 1rem;
+    right: 1rem;
+    background: none;
+    border: none;
+    color: white;
+    font-size: 1.5rem;
+    cursor: pointer;
+    padding: 0.5rem;
+}
+
+.mobile-menu-content {
+    margin-top: 10px;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+    padding: 1rem;
+}
+
+@media (max-width: 768px) {
+    .headbar {
+        flex-wrap: nowrap; /* Changed from wrap to nowrap */
+    }
+
+    .headbar-search {
+        order: 1;
+        flex: 1;
+        margin: 0 1rem;
+    }
+
+    .headbar-title {
+        order: 0;
+    }
+
+    .mobile-menu-button {
+        order: 2;
+        display: block;
+    }
 }
